activate hot module replacement
Hot module replacement is an advanced feature.
Updates will not full-reload the page, instead it will only update changed
parts and preserve states.

We need `react-hot-loader` for this.

```
yarn add react-hot-loader
```

It is safe to add it as a dependency, because it will not take effect in
production mode.

First we need to use its Babel plugin, so we add `react-hot-loader/babel`
to `.babelrc`.

Then we need to use `hot` to decorate the root component.

The latest step is to add `--hot` option to `webpack-dev-server`.
This option will automatically do the following things.

1. adds `webpack.HotModuleReplacementPlugin` to all entries
2. sets `devServer.hot` to `true` for all entries

You can also do these manually in `webpack.config.ts` if you want more control.
But remember **DO NOT** mix CLI option and `devServer` settings together.